refactor: Screen 구조 세분화 (검색 결과창 추출)
